requirements.txt
streamlit
pandas
numpy
transformers
torch
pillow
scikit-learn